salsa-ci: switch to using branch2repo
branch2repo introduces the ability to include additional packages into
the repository that it builds. We're taking advantage of that in order
to pull in tweaked versions of net-retriever and anna. Those in turn
give D-I the ability to consult an additional repository. Finally, that
allows the miniiso we've built to download the (unreleased) apt-setup
udeb we've built, so we can actually test it.